+++ test/functional/rpc_fundrawtransaction.py
@@ -390,7 +390,7 @@ class RawTransactionsTest(BitcoinTestFramework):
         addr2Obj = self.nodes[1].validateaddress(addr2)
 
         mSigObj = self.nodes[1].addmultisigaddress(
-            2, [addr1Obj['pubkey'], addr2Obj['pubkey']])
+            2, [addr1Obj['pubkey'], addr2Obj['pubkey']])['address']
 
         inputs = []
         outputs = {mSigObj: 1.1}
@@ -423,7 +423,7 @@ class RawTransactionsTest(BitcoinTestFramework):
         addr5Obj = self.nodes[1].validateaddress(addr5)
 
         mSigObj = self.nodes[1].addmultisigaddress(
-            4, [addr1Obj['pubkey'], addr2Obj['pubkey'], addr3Obj['pubkey'], addr4Obj['pubkey'], addr5Obj['pubkey']])
+            4, [addr1Obj['pubkey'], addr2Obj['pubkey'], addr3Obj['pubkey'], addr4Obj['pubkey'], addr5Obj['pubkey']])['address']
 
         inputs = []
         outputs = {mSigObj: 1.1}
@@ -450,7 +450,7 @@ class RawTransactionsTest(BitcoinTestFramework):
         addr2Obj = self.nodes[2].validateaddress(addr2)
 
         mSigObj = self.nodes[2].addmultisigaddress(
-            2, [addr1Obj['pubkey'], addr2Obj['pubkey']])
+            2, [addr1Obj['pubkey'], addr2Obj['pubkey']])['address']
 
         # send 1.2 BTC to msig addr
         txId = self.nodes[0].sendtoaddress(mSigObj, 1.2)