+++ doc/release-notes.md
@@ -3,4 +3,4 @@ Bitcoin ABC version 0.19.6 is now available from:
   <https://download.bitcoinabc.org/0.19.6/>
 
 This release includes the following features and fixes:
-
+ - Using addresses in createmultisig is now deprectated. Use -deprecatedrpc=createmultisig to get the old behavior.